requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
sentence-transformers==2.6.1
torch==2.2.2
faiss-cpu
transformers==4.40.1
tqdm==4.66.4
